BPMSPL-653: Separated task administration to new view
