+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/java/org/jbpm/workbench/ht/client/resources/i18n/Constants.java
@@ -32,6 +32,8 @@ public interface Constants extends Messages {
     Constants INSTANCE = GWT.create(Constants.class);
 
     String Tasks_List();
+    
+    String Tasks_Admin();
 
     String Grid();
