+++ jbpm-wb-human-tasks/jbpm-wb-human-tasks-client/src/main/resources/org/jbpm/workbench/ht/client/resources/i18n/Constants.properties
@@ -1,4 +1,5 @@
 Tasks_List=Tasks List
+Tasks_Admin=Task Administration List
 Actual_Owner=Actual Owner
 Reminder_Details=Reminder Details
 Reminder=Reminder
