+++ jbpm-wb-showcase/src/main/resources/org/jbpm/workbench/client/i18n/Constants.properties
@@ -1,4 +1,5 @@
 Tasks_List=Tasks List
+Tasks_Admin=Task Administration
 Process_Authoring=Process Authoring
 Authoring=Authoring
 Process_Management=Process Management
